+++ openeo/__init__.py
@@ -13,10 +13,9 @@ class BaseOpenEoException(Exception):
 
 from openeo._version import __version__
 from openeo.imagecollection import ImageCollection
-from openeo.rest.datacube import DataCube
+from openeo.rest.datacube import DataCube, UDF
 from openeo.rest.connection import connect, session, Connection
 from openeo.rest.job import BatchJob, RESTJob
-from openeo.internal.graph_building import UDF
 
 
 def client_version() -> str: