+++ openeo/rest/connection.py
@@ -13,7 +13,6 @@ from typing import Dict, List, Tuple, Union, Callable, Optional, Any, Iterator
 from urllib.parse import urljoin
 
 import requests
-from deprecated.sphinx import deprecated
 from requests import Response
 from requests.auth import HTTPBasicAuth, AuthBase
 
@@ -23,6 +22,7 @@ from openeo.config import get_config_option, config_log
 from openeo.internal.graph_building import PGNode, as_flat_graph
 from openeo.internal.jupyter import VisualDict, VisualList
 from openeo.internal.processes.builder import ProcessBuilderBase
+from openeo.internal.warnings import legacy_alias, deprecated
 from openeo.metadata import CollectionMetadata
 
 from openeo.rest import OpenEoClientException, OpenEoApiError, OpenEoRestError
@@ -38,7 +38,7 @@ from openeo.rest.job import BatchJob, RESTJob
 from openeo.rest.rest_capabilities import RESTCapabilities
 from openeo.rest.service import Service
 from openeo.rest.udp import RESTUserDefinedProcess, Parameter
-from openeo.util import ensure_list, legacy_alias, dict_no_none, rfc3339, load_json_resource, LazyLoadCache, \
+from openeo.util import ensure_list, dict_no_none, rfc3339, load_json_resource, LazyLoadCache, \
     ContextTimer, str_truncate
 from openeo.rest.localconnection import LocalConnection
 
@@ -959,6 +959,7 @@ class Connection(RestApiConnection):
             temporal_extent: Optional[List[Union[str, datetime.datetime, datetime.date]]] = None,
             bands: Optional[List[str]] = None,
             properties: Optional[Dict[str, Union[str, PGNode, Callable]]] = None,
+            max_cloud_cover: Optional[float] = None,
             fetch_metadata=True,
     ) -> DataCube:
         """
@@ -969,12 +970,17 @@ class Connection(RestApiConnection):
         :param temporal_extent: limit data to specified temporal interval
         :param bands: only add the specified bands
         :param properties: limit data by metadata property predicates
+        :param max_cloud_cover: shortcut to set maximum cloud cover ("eo:cloud_cover" collection property)
         :return: a datacube containing the requested data
+
+        .. versionadded:: 0.13.0
+            added the ``max_cloud_cover`` argument.
         """
         if self._api_version.at_least("1.0.0"):
             return DataCube.load_collection(
                 collection_id=collection_id, connection=self,
                 spatial_extent=spatial_extent, temporal_extent=temporal_extent, bands=bands, properties=properties,
+                max_cloud_cover=max_cloud_cover,
                 fetch_metadata=fetch_metadata,
             )
         else: